Move run_pager() and get_default_pager() from 'core' to 'output' module
