+++ osc/output/__init__.py
@@ -1,6 +1,8 @@
 from .key_value_table import KeyValueTable
 from .input import get_user_input
+from .output import get_default_pager
 from .output import print_msg
+from .output import run_pager
 from .output import sanitize_text
 from .output import safe_print
 from .output import safe_write
